api/requirements/base.txt
Django==3.2.6
django-extensions==3.1.3
django-cors-headers==3.8.0
django-environ==0.4.5
django-mysql==2.2.0
mysqlclient==2.1.0
dj-database-url==0.5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
python-decouple==3.6.0
gunicorn==20.1.0
drf-yasg==1.20.0